scrub sensitive values
