+++ internal/version/components.json
@@ -1,6 +1,6 @@
 {
   "config": "v0.40.0",
   "hosted-authenticate-oidc": "v0.1.0",
-  "mcp": "v0.33.0",
+  "mcp": "v0.34.0",
   "ssh": "v0.23.0"
 }
